fix: backfill NotifiedAt on legacy terminal runs before notification deploy
The backfill healing branch in process-update-laboratory-run.lambda.ts now
also heals a terminal run missing NotifiedAt (d079bb26). That branch is also
reached by the front-end's duration-backfill trigger whenever a user views
any pre-existing terminal run missing RunDurationSeconds — which is every
run that finished before this feature deployed. Without a pre-deploy
migration, the first time a user reopens an old run after deploy, the
healing branch would treat it as never-notified and send a stale "your run
finished" email for a run that may have finished months ago.

Extends the existing one-time backfill-poll-status.ts script (renamed to
backfill-laboratory-run-attributes.ts, since it now performs two independent
passes) with a second pass that stamps NotifiedAt onto every currently
terminal run missing it, so the healing branch finds nothing to heal for any
pre-existing run once this runs before deploy.
